fix: Simple bug fixes. * Fixes #145    * Renamed `intergration` to `integration` globally * Fixes #217    * Added documentation comments before `syntax`, `import`, `package`, `option`,      and `reserved` keywords in grammar       * The root cause is the change to docComment to make it non-skipped, so         we have to add it everywhere. Skipped comments (the `//` type) can be         absolutely anywhere because the grammar _skips_ them.    * Note, this _might_ also fix #319
