+++ pbj-integration-tests/src/test/java/com/hedera/pbj/integration/test/CompareToTest.java
@@ -1,4 +1,4 @@
-package com.hedera.pbj.intergration.test;
+package com.hedera.pbj.integration.test;
 
 import static java.util.Collections.shuffle;
 import static java.util.Collections.sort;
