+++ pbj-integration-tests/src/test/java/com/hedera/pbj/integration/test/ExtendedUtf8MessageWithStringTest.java
@@ -1,4 +1,4 @@
-package com.hedera.pbj.intergration.test;
+package com.hedera.pbj.integration.test;
 
 import com.google.protobuf.CodedOutputStream;
 import com.hedera.pbj.runtime.io.buffer.BufferedData;
