+++ pbj-integration-tests/src/test/java/com/hedera/pbj/integration/test/FieldsNonNullTest.java
@@ -1,4 +1,4 @@
-package com.hedera.pbj.intergration.test;
+package com.hedera.pbj.integration.test;
 
 import com.hedera.hapi.node.base.FeeSchedule;
 import com.hedera.hapi.node.base.TransactionFeeSchedule;
