+++ pbj-integration-tests/src/test/java/com/hedera/pbj/integration/test/JsonCodecTest.java
@@ -1,4 +1,4 @@
-package com.hedera.pbj.intergration.test;
+package com.hedera.pbj.integration.test;
 
 import com.google.protobuf.ByteString;
 import com.google.protobuf.util.JsonFormat;
