+++ pbj-integration-tests/src/test/java/com/hedera/pbj/integration/test/MalformedMessageTest.java
@@ -1,4 +1,4 @@
-package com.hedera.pbj.intergration.test;
+package com.hedera.pbj.integration.test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
