+++ pbj-integration-tests/src/test/java/com/hedera/pbj/integration/test/TestHashFunctions.java
@@ -1,20 +1,4 @@
-/*
- * Copyright (C) 2025 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hedera.pbj.intergration.test;
+package com.hedera.pbj.integration.test;
 
 import com.hedera.pbj.test.proto.pbj.Hasheval;
 import com.hedera.pbj.test.proto.pbj.TimestampTest;
@@ -27,8 +11,8 @@ import java.security.NoSuchAlgorithmException;
 public final class TestHashFunctions {
     public static int hash1(Hasheval hashEval) {
         try {
-            byte[] hash = MessageDigest.getInstance("SHA-256")
-                    .digest(Hasheval.PROTOBUF.toBytes(hashEval).toByteArray());
+            byte[] hash = MessageDigest.getInstance("SHA-256").digest(
+                    Hasheval.PROTOBUF.toBytes(hashEval).toByteArray());
             int res = hash[0] << 24 | hash[1] << 16 | hash[2] << 8 | hash[3];
             return processForBetterDistribution(res);
         } catch (NoSuchAlgorithmException e) {
@@ -84,10 +68,10 @@ public final class TestHashFunctions {
         }
         if (hashEval.subObject() != Hasheval.DEFAULT.subObject()) {
             TimestampTest sub = hashEval.subObject();
-            if (sub.nanos() != TimestampTest.DEFAULT.nanos()) {
+            if (sub.nanos() != sub.DEFAULT.nanos()) {
                 result = 31 * result + Integer.hashCode(sub.nanos());
             }
-            if (sub.seconds() != TimestampTest.DEFAULT.seconds()) {
+            if (sub.seconds() != sub.DEFAULT.seconds()) {
                 result = 31 * result + Long.hashCode(sub.seconds());
             }
         }
@@ -95,8 +79,7 @@ public final class TestHashFunctions {
             result = 31 * result + hashEval.text().hashCode();
         }
         if (hashEval.bytesField() != Hasheval.DEFAULT.bytesField()) {
-            result = 31 * result
-                    + (hashEval.bytesField() == null ? 0 : hashEval.bytesField().hashCode());
+            result = 31 * result + (hashEval.bytesField() == null ? 0 : hashEval.bytesField().hashCode());
         }
 
         return processForBetterDistribution(result);
